requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.0
pydantic==2.9.2
python-multipart==0.0.9  # Necessário para UploadFile/FormData
mysql-connector-python==9.1.0  # Conector MySQL usado em cadastro.py
pydub==0.25.1  # Conversão de áudio para formato WAV
azure-cognitiveservices-speech  # Azure Speech SDK para avaliação de pronúncia
python-dotenv>=1.0.0  # Carregar variáveis de ambiente de um arquivo .env

# Dependências opcionais (descomente se quiser utilizar/tests)
# httpx==0.27.2  # Cliente HTTP para chamadas externas / testes
# pytest==8.2.0  # Testes automatizados
